requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
dj-rest-auth==4.0.0
Django==4.0
django-ckeditor==6.5.1
django-cors-headers==3.10.0
django-environ==0.9.0
django-js-asset==2.0.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
environs==9.5.0
marshmallow==3.19.0
packaging==23.1
Pillow==9.5.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
whitenoise==5.3.0
